Updated data set processing scripts and configuration files, includin…

+++ Code/dataloader/CIFAR10.py
@@ -68,7 +68,7 @@ class CIFAR10(Dataset):
 
         # doing this so that it is consistent with all other datasets
         # to return a PIL Image
-        img = Image.fromarray(img)
+        img = Image.fromarray(np.uint8(img))
 
         if self.transform is not None:
             img = self.transform(img)
